refactor(ff-filter): move FilterGraph impl from graph/mod.rs to graph/graph.rs
